+++ include/openthread/BUILD.gn
@@ -43,6 +43,7 @@ source_set("openthread") {
     "backbone_router_ftd.h",
     "border_agent.h",
     "border_router.h",
+    "border_routing.h",
     "channel_manager.h",
     "channel_monitor.h",
     "child_supervision.h",
