+++ include/openthread/instance.h
@@ -53,7 +53,7 @@ extern "C" {
  * @note This number versions both OpenThread platform and user APIs.
  *
  */
-#define OPENTHREAD_API_VERSION (223)
+#define OPENTHREAD_API_VERSION (224)
 
 /**
  * @addtogroup api-instance
