+++ src/core/BUILD.gn
@@ -308,6 +308,7 @@ openthread_core_files = [
   "api/backbone_router_ftd_api.cpp",
   "api/border_agent_api.cpp",
   "api/border_router_api.cpp",
+  "api/border_routing_api.cpp",
   "api/channel_manager_api.cpp",
   "api/channel_monitor_api.cpp",
   "api/child_supervision_api.cpp",
